requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
opencv-python==4.8.1.78
face_recognition==1.3.0
numpy==1.26.2
werkzeug==3.0.1
